refactor: 예외 응답을 응답 디렉터리로 이동
