+++ src/main/java/ex/sample/global/response/InvalidInputRes.java
@@ -1,4 +1,4 @@
-package ex.sample.global.exception;
+package ex.sample.global.response;
 
 public record InvalidInputRes(
     String field,
